Уведомление о начале события
конфликты, фиксы
